Updates for 2.0.5 updated release notes
A320 demo slightly modified for normal use so that it does not
load extra cloud layers that are not used except during our looping
IITSEC demo to save time during startup of the demo.

Slightly modified anti-collision lights on A320 so that show light
patterns on the runway when the A320 is on the ground which looks a
bit more like they would in reality.